pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specfoundry"
version = "0.1.0"
description = "Autonomous specification generator for LLM-driven development"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
specfoundry = "specfoundry.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["specfoundry*"]
